+++ src/components/CharacterForm.js
@@ -10,7 +10,7 @@ import Select from '@material-ui/core/Select';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Loader from "./Loader";
-import RequestService from "./RequestService";
+import RequestService from "../services/RequestService";
 import CharacterInfos from "./CharacterInfos";
